Plane are now flying correctly
I had to use export / import statements for shared variables, because the scripts are declared as modules, so that es6 and anime.js can work correctly. As a result some variables had to be moved between files, so that they are declared there where they get changed. You can only read in other files and not change the variable.